+++ lib/src/result/failure.dart
@@ -1,5 +1,9 @@
 part of tie_fp;
 
+/// `Failure` descends from [Result]
+/// allowing to safely return wrap an [Exception]
+/// and get the [StackStrace] if applies.
+
 final class Failure<T> extends Result<T> {
   final Object? error;
 
@@ -11,21 +15,21 @@ final class Failure<T> extends Result<T> {
   }
 
   @override
-  String getError() {
-    return error.toString();
-  }
+  Object? getError() => error;
 
   @override
-  T? getValue() => null;
+  String toString() => error.toString();
 
   @override
   bool isError() => true;
-
-  @override
-  Object? getRawError() => error;
 
   final StackTrace? _stackTrace;
 
   @override
   StackTrace? stackTrace() => _stackTrace;
+
+  @override
+  T getValue() {
+    throw Failure('failure data cannot return a value');
+  }
 }