+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.5
+
+- Add documentation.
+- Remove nullable from `getValue`.
+- Add `Result.wrapFunction()` factory.
+- Remove `T Function()` extension.
+
 ## 0.0.4
 
 - Add `onError` callback to unified all error messages.